# openapi/evry-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "currentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The user profile no longer exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "updateCurrentUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatedUser"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The user profile no longer exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/readiness/checkin": {
      "post": {
        "operationId": "ReadinessController_checkin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadinessInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Estado diario creado o actualizado para la fecha civil de hoy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "readiness"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/readiness/latest": {
      "get": {
        "operationId": "ReadinessController_latest",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Estado de la fecha civil de hoy, o null si todavía no se registró.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "nullable": true,
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Readiness"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "readiness"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/routines": {
      "get": {
        "operationId": "RoutinesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Routine"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "RoutinesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoutineDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Routine"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/routines/{id}": {
      "get": {
        "operationId": "RoutinesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Routine"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "RoutinesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoutineDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Routine"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "RoutinesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/routines/{id}/start": {
      "post": {
        "operationId": "RoutinesController_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Routines"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EVRY API",
    "description": "API privada de seguimiento de fuerza EVRY",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "evry_refresh"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "biologicalSex": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_SAY"
            ],
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "STRENGTH",
                "HYPERTROPHY",
                "ENDURANCE",
                "FAT_LOSS",
                "GENERAL_FITNESS",
                "MOBILITY"
              ]
            }
          },
          "trackCycle": {
            "type": "boolean"
          },
          "avgCycleLen": {
            "type": "integer"
          },
          "avgPeriodLen": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "biologicalSex",
          "birthDate",
          "goals",
          "trackCycle",
          "avgCycleLen",
          "avgPeriodLen",
          "createdAt"
        ]
      },
      "UserUpdateInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "biologicalSex": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_SAY"
            ],
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 date or date-time. Omitted or null values leave the current birth date unchanged."
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "STRENGTH",
                "HYPERTROPHY",
                "ENDURANCE",
                "FAT_LOSS",
                "GENERAL_FITNESS",
                "MOBILITY"
              ]
            }
          },
          "trackCycle": {
            "type": "boolean"
          },
          "avgCycleLen": {
            "type": "integer",
            "minimum": 20,
            "maximum": 45
          },
          "avgPeriodLen": {
            "type": "integer",
            "minimum": 2,
            "maximum": 10
          }
        }
      },
      "UpdatedUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "biologicalSex": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_SAY"
            ],
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "STRENGTH",
                "HYPERTROPHY",
                "ENDURANCE",
                "FAT_LOSS",
                "GENERAL_FITNESS",
                "MOBILITY"
              ]
            }
          },
          "trackCycle": {
            "type": "boolean"
          },
          "avgCycleLen": {
            "type": "integer"
          },
          "avgPeriodLen": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "biologicalSex",
          "birthDate",
          "goals",
          "trackCycle",
          "avgCycleLen",
          "avgPeriodLen"
        ]
      },
...
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "Routine": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "dayOfWeek": {
            "type": "integer",
            "nullable": true,
            "minimum": 0,
            "maximum": 6
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoutineExercise"
            }
          }
        },
        "required": [
          "id",
          "userId",
          "name",
          "dayOfWeek",
          "notes",
          "createdAt",
          "updatedAt",
          "exercises"
        ]
      },
      "Workout": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "cancelledAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "COMPLETED",
              "CANCELLED"
            ],
            "type": "string"
          },
          "clientId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "lastSyncId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "revision": {
            "type": "integer"
          },
          "cyclePhase": {
            "nullable": true,
            "enum": [
              "MENSTRUAL",
              "FOLLICULAR",
              "OVULATION",
              "LUTEAL"
            ],
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "routineId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutSet"
            }
          },
          "routine": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Routine"
              }
            ]
          }
        },
        "required": [
          "id",
          "userId",
          "name",
          "startedAt",
          "endedAt",
          "cancelledAt",
          "status",
          "clientId",
          "lastSyncId",
          "revision",
          "cyclePhase",
          "notes",
          "routineId",
          "createdAt",
          "updatedAt",
          "sets",
          "routine"
        ]
      },
...
      "Readiness": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "civilDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "Fecha civil persistida y serializada como medianoche UTC; puede ser null en registros anteriores."
          },
          "sleepHrs": {
            "type": "number",
            "nullable": true
          },
          "stress": {
            "type": "integer",
            "nullable": true
          },
          "soreness": {
            "type": "integer",
            "nullable": true
          },
          "motivation": {
            "type": "integer",
            "nullable": true
          },
          "score": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          }
        },
        "required": [
          "id",
          "userId",
          "date",
          "civilDate",
          "sleepHrs",
          "stress",
          "soreness",
          "motivation",
          "score"
        ]
      },
      "ReadinessInput": {
        "type": "object",
        "properties": {
          "sleepHrs": {
            "type": "number",
            "minimum": 0,
            "maximum": 16,
            "nullable": true
          },
          "stress": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "nullable": true
          },
          "soreness": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "nullable": true
          },
          "motivation": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "nullable": true
          }
        }
      },
...
      "RoutineExerciseDto": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "string"
          },
          "order": {
            "type": "number",
            "minimum": 0
          },
          "targetSets": {
            "type": "number",
            "minimum": 1,
            "maximum": 20
          },
          "targetReps": {
            "type": "number",
            "minimum": 1,
            "maximum": 100
          },
          "targetWeightKg": {
            "type": "number",
            "minimum": 0
          },
          "seriesPlan": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoutineSeriesPlanDto"
            }
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "exerciseId",
          "order",
          "targetSets"
        ]
      },
      "CreateRoutineDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "dayOfWeek": {
            "type": "number",
            "minimum": 0,
            "maximum": 6
          },
          "notes": {
            "type": "string"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoutineExerciseDto"
            }
          }
        },
        "required": [
          "name",
          "exercises"
        ]
      },
      "UpdateRoutineDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "dayOfWeek": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 6
          },
          "notes": {
            "type": "string"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoutineExerciseDto"
            }
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "retryable": {
            "type": "boolean"
          },
          "requestId": {
            "type": "string"
          }
        }
      }
    }
  }
}